requirements/all.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    make upgrade
#
-e git+https://github.com/jsocol/django-ratelimit.git@72edbe8949fbf6699848e5847645a1998f121d46#egg=ratelimit
analytics-python==1.2.9
argparse==1.4.0
astroid==2.3.3
attrs==17.4.0
bcrypt==3.1.7
bok-choy==1.0.1
boto==2.49.0
cached-property==1.5.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click-log==0.3.2
click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cryptography==2.8
ddt==1.2.2
defusedxml==0.6.0
django-appconf==1.0.3
django-compat==1.0.15
django-debug-toolbar==1.6
django-extensions==2.2.8
django-filter==1.0.4
django-hijack==2.1.10
django-rest-swagger==2.2.0
django-ses==0.8.14
django-sortedm2m==3.0.0
django-statici18n==1.9.0
django-storages==1.9.1
django-waffle==0.19.0
django-webpack-loader==0.6.0
django==1.11.28
djangorestframework-jwt==1.11.0
djangorestframework==3.11.0
docker-compose==1.25.4
docker[ssh]==4.2.0
dockerpty==0.4.1
docopt==0.6.2
edx-ace==0.1.12
edx-auth-backends==3.0.2
edx-django-release-util==0.3.6
edx-django-sites-extensions==2.4.3
edx-django-utils==2.0.4
edx-drf-extensions==2.4.6
edx-i18n-tools==0.5.0
edx-lint==1.4.1
edx-opaque-keys==2.0.1
edx-rest-api-client==3.0.2
git+https://github.com/edx/credentials-themes.git@0.1.30#egg=edx_credentials_themes==0.1.30
factory-boy==2.12.0
faker==4.0.0
future==0.18.2
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
httpretty==0.9.7
idna==2.7
importlib-metadata==1.5.0
isort==4.2.5
itypes==1.1.0
jinja2==2.11.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lazy==1.4
markdown==3.2
markupsafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==8.2.0
mysqlclient==1.4.6
newrelic==5.6.0.135
nodeenv==1.3.5
oauthlib==3.1.0
olefile==0.46
openapi-codec==1.3.2
packaging==20.1
paramiko==2.7.1
path.py==12.0.2
pathlib2==2.3.5
pbr==5.4.4
pep8==1.7.1
pillow==4.0.0
pluggy==0.13.1
polib==1.1.0
psutil==1.2.1
py==1.8.1
pycparser==2.19
pycryptodomex==3.9.6
pygments==2.5.2
pyinotify==0.9.6 ; "linux" in sys_platform
pyjwkest==1.3.2
pyjwt==1.7.1
pylint-celery==0.3
pylint-django==2.0.11
pylint-plugin-utils==0.6
pylint==2.4.2
pymongo==3.10.1
pynacl==1.3.0
pyparsing==2.4.6
pyrsistent==0.15.7
pytest-django==3.8.0
pytest==5.3.5
python-dateutil==2.8.1
python-memcached==1.59
python-slugify==1.2.6
python3-openid==3.1.0
pytz==2019.3
pyyaml==3.12
requests-oauthlib==1.3.0
requests==2.20.1
responses==0.10.9
rest-condition==1.0.3
sailthru-client==2.2.3
selenium==3.141.0
semantic-version==2.8.4
simplejson==3.17.0
six==1.14.0
slumber==0.7.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
stevedore==1.10.0
testfixtures==6.12.0
text-unidecode==1.3
texttable==1.6.2
transifex-client==0.13.7
typed-ast==1.4.1
unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.8
websocket-client==0.57.0
wrapt==1.11.2
xss-utils==0.1.2
zipp==1.1.0

# The following packages are considered to be unsafe in a requirements file:
# setuptools
